refactoring
* merged JcrWriter and JcrPersistImpl into a single service
* added @NotNull jetbrains annotations
