+++ SlingJCRPersist/pom.xml
@@ -150,6 +150,12 @@
             <version>2.0</version>
             <type>jar</type>
         </dependency>
+        <dependency>
+            <groupId>org.jetbrains</groupId>
+            <artifactId>annotations</artifactId>
+            <version>16.0.2</version>
+            <scope>provided</scope>
+        </dependency>
     </dependencies>
 
     <build>
